refactor: 페이지 레이아웃과 실제 컴포넌트 파일 분리
