+++ src/pages/addExpense/AddExpensePage.tsx
@@ -2,70 +2,20 @@
 import styled from 'styled-components';
 import { flexCenter, flexColumnCenter } from '@styles/CommonStyles';
 
-import TopNavigation from '@layout/TopNavigation';
 import type { ExpenseFormType } from '@models/expense';
 
 import { useForm, FormProvider } from 'react-hook-form';
-import { useNavigate } from 'react-router-dom';
 import { useState } from 'react';
 
 import useIsDemoMode from '@hooks/useIsDemo';
 import useSaveExpense from './hooks/useSaveExpense';
 
+import NavigationLayout from './navigation';
 import WriteExpense from './components/WriteExpense';
 import WriteEmotion from './components/WriteEmotion';
 import WriteSatisfaction from './components/WriteSatisfaction';
 
-import MetaThemeColor from '@components/background/MetaThemeColor';
 import LoadingModal from '@components/modal/LoadingModal';
-
-type AddNavProps = {
-  children: React.ReactNode;
-  title: string;
-  isDemoMode: boolean;
-  hasPrev: boolean;
-  prevStep: () => void;
-};
-
-const NavigationLayout = ({ children, title, isDemoMode, hasPrev, prevStep }: AddNavProps) => {
-  const navigate = useNavigate();
-  return (
-    <>
-      <MetaThemeColor color="#F4F4F4" />
-      <TopNavigation
-        _TopBar={
-          <TopNavigation.TopBar
-            leftContent={
-              hasPrev && (
-                <TopNavigation.TopBar.PrevButton
-                  onClick={() => {
-                    prevStep();
-                  }}
-                />
-              )
-            }
-            centerContent={
-              <TopNavigation.TopBar.CenterTitle>
-                {title}
-                {isDemoMode && (
-                  <span style={{ fontSize: '12px', color: '#47cfb0' }}> (체험중)</span>
-                )}
-              </TopNavigation.TopBar.CenterTitle>
-            }
-            rightContent={
-              <TopNavigation.TopBar.CloseButton
-                onClick={() => {
-                  navigate(-1);
-                }}
-              />
-            }
-          />
-        }
-      />
-      {children}
-    </>
-  );
-};
 
 const AddExpensePage = () => {
   const isDemoMode = useIsDemoMode();
