+++ src/pages/dashboard/components/EmotionList.tsx
@@ -1,7 +1,7 @@
 import styled from 'styled-components';
 import { divider, flexBetween, flexCenter, flexColumnCenter } from '@styles/CommonStyles';
 
-import type { EmotionAmountTotalType } from '@models/api/dashboard';
+import type { EmotionAmountTotalType } from '@models/service/dashboard';
 import { addCommasToNumber, calculatePercentages } from '@utils/index';
 import { getEmotionIcon, getEmotionText } from '@models/emotion';
