+++ src/pages/expenseDetailView/ExpenseDetailViewPage.tsx
@@ -1,4 +1,4 @@
-import styled, { css } from 'styled-components';
+import styled from 'styled-components';
 import {
   flexCenter,
   flexColumnCenter,
@@ -6,7 +6,6 @@ import {
   overflowWithoutScroll,
   summaryArea,
 } from '@styles/CommonStyles';
-import TopNavigation from '@layout/TopNavigation';
 
 import { useNavigate } from 'react-router-dom';
 
@@ -22,6 +21,7 @@ import useAICommentData from './hooks/useAICommentData';
 import type { EmotionKey, Register } from '@models/index';
 import type { ExpenseFormType } from '@models/expense';
 
+import NavigationLayout from './navigation';
 import ExpenseSummary from '@components/expense/ExpenseSummary';
 import MultiText from '@components/input/MultiText';
 import LoadingModal from '@components/modal/LoadingModal';
@@ -29,101 +29,8 @@ import { VolumeBtn } from '@components/button';
 import EditSummary from './components/EditSummary';
 
 import { formatAmountNumber, getSpendSumamryText } from '@utils/index';
-import MetaThemeColor from '@components/background/MetaThemeColor';
 import Spinner from '@components/information/Spinner';
 import { PagePath } from '@models/navigation';
-
-type NavLayoutProps = {
-  children: React.ReactNode;
-  isEdit: boolean;
-  isValid: boolean;
-  handleSubmit: () => void;
-  handleDelete: () => void;
-  handleMovePrevPage: () => void;
-  toggleEdit: () => void;
-};
-
-const NavigationLayout = ({
-  children,
-  isEdit,
-  isValid,
-  handleSubmit,
-  handleDelete,
-  handleMovePrevPage,
-  toggleEdit,
-}: NavLayoutProps) => {
-  return (
-    <>
-      <MetaThemeColor color="#F4F4F4" />
-      <TopNavigation
-        _TopBar={
-          <TopNavigation.TopBar
-            leftContent={
-              !isEdit && <TopNavigation.TopBar.PrevButton onClick={handleMovePrevPage} />
-            }
-            centerContent={
-              <TopNavigation.TopBar.CenterTitle>작성완료 내역</TopNavigation.TopBar.CenterTitle>
-            }
-            rightContent={
-              <Toolbar>
-                {isEdit ? (
-                  <SaveButton onClick={handleSubmit} disabled={isValid}>
-                    완료
-                  </SaveButton>
-                ) : (
-                  <EditButton onClick={toggleEdit} />
-                )}
-                <DeleteButton onClick={handleDelete} />
-              </Toolbar>
-            }
-          />
-        }
-      />
-      {children}
-    </>
-  );
-};
-
-const Toolbar = styled.div`
-  ${flexCenter}
-  gap: 16px;
-`;
-
-const toolbarStyle = css`
-  width: 24px;
-  height: 24px;
-  color: #bcbcbc;
-  cursor: pointer;
-
-  &:hover {
-    color: #47cfb0;
-  }
-`;
-
-const EditButton = styled(TopNavigation.TopBar.EditButton)`
-  ${toolbarStyle}
-`;
-
-const DeleteButton = styled(TopNavigation.TopBar.DeleteButton)`
-  ${toolbarStyle}
-`;
-
-const SaveButton = styled.button`
-  color: #47cfb0;
-  font-size: 16px;
-  font-weight: 700;
-  &:hover {
-    filter: brightness(1.1);
-  }
-
-  &:disabled {
-    color: #ccc;
-    cursor: not-allowed;
-    &:hover {
-      filter: none;
-    }
-  }
-`;
 
 const ExpenseDetailViewPage = () => {
   const { id: paramId } = useParams();
