+++ src/pages/statistics/components/Satisfaction/hooks/useSatisfactionData.tsx
@@ -1,4 +1,5 @@
 import { useQuery } from 'react-query';
+
 import {
   fetchSatisfactionByGender,
   fetchSatisfactionByMbti,
